Added full CRUD methods for boards and posts
This includes the update and delete methods but also a few other helper
methods for threads. User can now pull a single thread object if they
need to. Threads also have a title in their metadata to optimize showing
all the threads for one board. Updating the first post in a thread also
updates the thread's title if it has changed. Deleting the first post in
a thread will update the thread's title by prefixing it with a "DELETED:
" string.

There is a few validations and some error checking but there needs to be
more in the future.