Getting threads per board now return a title
The handler that wraps the createReadStream that gets all the threads
for a board was modified so that for each thread it gets, it goes and
find the first post of that threads and attaches the title of the post
to the thread object being returned.

The make_posts test file was modified to include a title in the post
object when creating posts.