Import methods
Added import methods for both boards and posts. Since importing old data
could bring with it older keys, an index needed to be created to match
the old keys to the new keys generated on import. Sublevel needed to be
added to the project to section off the index that matched the old keys
to the new keys.

The import methods having defined a new index, the delete methods for
both boards and post needed to be updated to clear the new index when
items are deleted. The deleteItr function was updated to clearly spell
out what keys needed to be created based on the type of post being
deleted.

New query methods were also created that would take in an old key and
return the matching new key for that object.

New tests concerning import were also created as well a new tear down
section in the last test to ensure no data is left in the db after the
tests have completed.