Refactor boards model
Refactored the boards model to be more modular by removing all input
validation into a separate file and using Joi to do our validation. Also
modified how the boards object in being exported. Now validation is run
on each function tied to board before actually calling the appropriate
function. If validation fails, the callback is called with the
validation error instead of the respective function.

Added path and updated most of the requires to use path instead of
concating the strings.