Threads/Posts refactor, bug fixes
THIS IS A BREAKING CHANGE. THE API NEEDS TO BE UPDATED.

Refactored the posts.js on root by splitting out all the thread related
code into it's own directory. Posts also has it's own directory. Each
directory has a main interface that is defined by the type of code it's
related to (posts.js/threads.js) and a validator.js file in each
directory. This validator.js just runs validation on all inputs for all
interface functions to ensure all inputs are correct, otherwise the
original callback is invoked with the list of errors.

Refactored both test files to point to the correct files as well as
update them to use the correct threads functions rather than rely on the
functions that used to be in posts. The seed files were updated in a
similar manner.

The boards code was also updated to fix a bug where the validator was
used to fill in the created_at property for all boards, but the
validator returns a string time and not a number of milliseconds since
epoch time. This has been fixed.

Other files were updated to use path.join instead of string
concatenation.

Note:
Although there are update and delete methods for threads. They should
not be used until their intentions and usage are reviewed more
thoroughly.

This code could be refactor even more.