Added board model to core boards crud
Started off by cleaning up the db references so that they are more
modular and readable. Also re-routed the index to point to the new board
controller file instead of the old board dir.

In the board model, validations were properly addressed. As well as, a
possible fix for catching promise error on validation. Also, the
getKeyFromId and getLegacyKeyFromId static methods were added and a
.prefix property that returns the prefix associated with this model was
added.

Tests were updated and cleaned of any extra console logs.