Boards model/db/schema/index overhaul
Boards have been overhauled so that the /boards/ dir is a one stop shop
for anything board related. There is a model that declares and
instantiates board models to be used by the db. A db file that takes
care of how board models are written to the db. A schema file that
handle board model validation, and a index file that now takes in boards
as json and converts them into board model for the db to process and
also converts it back to json on return.

Some additional files were edited to help this process. Additional
prototype methods were added to the board model to help validate the
board and to return board model objects as json objects. The db and the
board index files were modified to use the new way boards are
instantiated and validated. Additionally, the create method in the db
file now handles generating boardIds right before insertion into the db.
The helper file was modified to optionally take in a timestamp but to
also use Date.now() if no timestamp is given. The board constructor
takes in a id but this id may be overwritten if it is a new board to be
inserted into the db.