Removed versions from board, board property fixes
Removed versioning from the board model.

Updated how the board model handles property management. Instead of
setting all the properties of a board to undefinded if the property is
not passed in, the property isn't created at all. The same is done for
the toObject method. If a property isn't set in the model, it isn't
created in the returning object.

Doing so broke a few tests and methods. Originally, it was assumed that
children would always return an array. It would be empty if there were
no children. Now it doesn't even set the property if there are no
children.