Thread Model test file
Created a Thread model test file to fully test all the methods and
properties of a thread model. This test file implements full coverage,
BDD testing, and promises to test the thread model.

Also adjusted the thread's model.js file to decouple the created_at and
updated_at properties from the id property in the constructor. This
keeps the created from being overridden on create if a created_at time
already exists (in the case of imported threads).