Changed post validation
Post validation now has two options, create and import. The difference
being that import is less strict on the user_id by not requiring it.

All other post related files were updated to account for this changes.
Posts schema was renamed to validate, Model and Index were edited to
reflect the name change. And the model test file was updated as well.