Bug fixes: board/post/thread legacy key, metadata
Fixed an issue where the board/post/thread legacy keys were not
being purged during their respective purge methods.

Also fixed a bug where during purge for threads, it was deleting the
thread username key even though it was the first post creating that key.
This causes issues whenever a thread is purged but the first post is
not. That piece of code is removed for now until how to handle purging
the first thread is figured out.