Tests for key management
Added a new test suite to test for all keys generated during all model
actions. This essentially ensure that as models are being created and
destroyed, all the associated keys are being properly created,
manipulated, and destroy respectively.

This test covers all 4 (currently) models: users, posts, thread, and
boards.

A new tool was needed to manually look into the DB and see which objects
were being created. This new tool, probe, basically reads and destroys
objects from the database.

As of writing this commit, the users model do not have much to test and
tests in general for this model are lacking. Also, the first post of a
thread is not being handled well.