Post index - Numbered index based
The Thread - Post index that lists all the posts within a thread has
been modified to use a numbered index instead of created_at time to keep
track of posts. This was mainly done to make pagination easier. The
byThread method has been modified to only take in limit and page as
options which define how many post per page, and which page to view
respectively.

The only other major change is an addition of a metadata field called
post_order that simply ties a post.id to the position of that post in a
thread. This was created as a quick lookup to purge a post and all of
it's properties when necessary.

Bytewise was added to encode the post index number for proper key
lookup.

All tests have been modified to account for this change.