Updated thread index to use last updated at
Updated the thread/post code so that each new post updates the board -
thread index with the most recent activity. In essence, the index is now
sorted by last updated time. Each new post removes the old updated at
time index for a thread, updates the lastPostCreatedAtKey, then adds a
new index object with the new update time.

The initial thread creation creates both the metadata property and the
index with the thread's created_at as a quick plant so that all
subsequent calls to posts.create has something to call without erroring.

The purge method for threads was also updated to ensure the index and
metadata are removed in the correct order.

This means that the updated_at property for threads are no longer need
and has been removed from all the thread code.

All tests have been updated to reflect the above change.