Fixed post/thread count for boards with children
Fixed an issue where boards with children did not have an accurate
post/thread count. Originally it did not take into count the amount of
posts/threads that any child boards had. Now it sums up all the
posts/threads that the parent board and all children board have.

The change was made in the Boards.find method but could be properly
addressed with more metadata changes.