Updated User code and updated tests
Added update, delete, and purge methods to users. The methods are also
used in the index.js file for users and the model itself was updated to
include a validateUpdate method and some key related methods. Full
coverage tests were also written to cover all of users as it is now.

Fixed a bug in boards.db where the import wasn't properly returning a
promise. It does so now.

Also Fixed a bug in boards testing that wasn't properly looking up the
board after it was undeleted.

Also updated the rest of the codebase to be more in line with each
other, the fixes include:
- All db update functions can now handle individual property updates.
- Proper naming of all vars in the index.js files
- Proper handling of the delete boolean in update

All tests were also updated to be more in line with each other, the
fixes include:
- Switching update_at checks to use .above where possible.
- Properly cleaning the DB after every test suite.
- Added a 'clean' method to probe that clears all DBs.

All tests pass.