Added a total post count and total thread count to boards
Added two new metadata keys to boards, total_post_count and
total_thread_count that keeps watch of how many posts and threads are
attached to each board. The difference between the total version vs the
regular version is that the total version keeps count of all posts and
thread of the current board and all children boards.

This was accomplished by adding a new (inc/dec)Total(post/thread)Count
methods to boards that mainly just increment or decrement the total
post/thread counts for the current board. The regular
inc/dec(post/thread)Count method was also modified so that after the
regular post/thread count is incremented, it also increments the total
post/thread count of the current board, and then calls the parent board
to increment it's total post/thread count. The effects propagate up
because the total version of the methods also calls its parent to update
the totat post/thread counts as well.

All tests were updated for this change and are passing.