Last post metadata optimization
Updated all last post related metadata in the post create method. The
issue was any new post with an old created_at property would be seen as
a new post in that thread. Which in turn would update the board's last
post metadata. Now, there's a check to see if the incoming post's
created_at is older than the current last post created at value. If it
is older, then it is ignored.

This change is more of a safeguard than a bug fix.