Board DB optimizations
Used Promise.join and batched calls to optimize some of the code in
board's db.js file. Most of the optimizations are low hanging fruit with
most of the fixes being nothing more than shortening the code to lessen
surface area for bugs. Other optimizations are more performance related
using join and batch to make a series of calls parallel instead. There
is room for more improvement.