Fixed old posts shown as last updated thread in boards
Fixed an issue where an old imported post with an obsolete created_at
property could end up as being shown as the last updated thread on the
homepage. The fix basically involves checking the board's current last
updated at timestamp and seeing if the incoming post's created_at
property is equal to or larger than that.