Threads test failing / byBoard bug
There was a bug in threads.byBoard where the first index was not being
returned. The issue lies is how the startKey was defined. It was
skipping over the first key because the suffix '\x00' was added to the
end of the key. This is not necessary if the start of the query is an
index position.