BBCode Pre processing, modified post body processing
Moved all BBCode parsing into the index.js file of posts.

Modified how body and encodedBody is being handled. Instead of saving
both properties as separate and non-synched values, post.body is now the
only resource that is checked upon import, create, and update. If there
is valid BBCode in body, the initial value of post.body is saved to
post.encodedBody, while post.body is saved with the parsed version of
itself.

Currently this post body processing is being moved to a file called
pre.js that will hold all pre-processing work before an object is sent
to the db. Any post processing work that needs to be done after being
sent to the db should be written in a post.js file and required from the
index.js file.