Cleaned up seed, fixed/changed testDB location
Cleaned up the code in seed and how it generates content. It is now
promise based and does not rely on async at all. It also has switched to
using the core wrapper rather than directly using db. The way seed is
called has been simplified to one method call rather than exposing an
interface which suits how we're using it currently.

The error that was causing the test files to run on the default db path
has been addressed and now run tests in a .testDB dir.