Optimized the use of Joi on Board
Optimized how Joi was being used as a validator on Boards. This includes
making use of the return of the validate function by piping it into the
db method it validates for. Also, the stripUnknown option was turned on
for the create and update validations so any extraneous data is stripped
out. Each method that takes in an input of some sort now has a
validation method in front of it.