Updated post validations, return user id on byThreads
Reverted a change where user.id was not returned on byThreads. This is
again being returned since it is being used to check if a post is
editable by the user.

The encodedBody has been removed from all post validation because it
does not matter at this point and should be stripped out. It will be
replaced by whatever body has or if there's no bbCode, it is just
ignored.