Post now only respects encodedBody
The post now ignores all content in body. It now accepts only the
encodedBody property. It also checks this property for any BBcode and
parses it if it exists. It then updates the post.body with the parsed
text.

If there is not BBCode, then the encodedBody is merely copied over and
the encodedBody is deleted.

All tests have been updated to reflect this change.